refactor: move AgentEvent to agent_loop module
Relocate the AgentEvent enum from channels/web/sse.rs to a new
agent_loop/event.rs module. The agent loop concept should not live
under a specific channel's namespace; channels are passive consumers
of these lifecycle events. web/sse.rs now re-exports the type so the
existing SSE payload shape is preserved with no wire-format change.
